+++ Broken/__init__.py
@@ -103,6 +103,7 @@ from Broken.Core.BrokenUtils import (
     Ignore,
     LazyImport,
     OnceTracker,
+    PlainTracker,
     SameTracker,
 )